clean up validator.js
