+++ server/lib/commonlib.js
@@ -5,7 +5,7 @@ var utf8 = require("utf8");
 
 var common = {};
 
-common.encoded = function(text){    
+common.encoded = function(text){
     var bytes = utf8.encode(text);
     var encoded = base64.encode(bytes); 
     return encoded;
@@ -27,6 +27,30 @@ common.errorHandler = function (error) {
     }
     console.log(_error);
 	return _error;
-}
+};
+
+common.validateRating = function (rating) {
+    var _rating = parseInt(rating);
+    if (_rating > 0 && _rating < 10) return true;
+    else return false;
+};
+
+common.validateEmail = function (email) {
+    var regex = /\S+@\S+\.\S+/;
+    if (regex.test(email)) return true;
+    else return false;
+};
+
+common.validateNumber = function (value) {
+    return true;
+};
+
+common.validateDate = function (value) {
+    return true;
+};
+
+common.validateString = function (value) {
+    return true;
+};
 
 module.exports = common;
